refactor: extract getTotalPurchasingPower to utils for better testability
